requirements.txt
fastapi==0.112.2
python-telegram-bot==13.7
pillow==10.4.0
pyzbar==0.1.9
uvicorn==0.30.6